refactor: urlRequest의 생성 방식을 결정하는 URLRequestTask 타입 구현
* URLRequstBuilder와 RequestType 의 로직을 URLRequestTask로 이동
* 재사용성을 위하여 NetworkManager request 메서드의 파라미터로 header와 body를 받도록 수정
