--- WeatherForecast/WeatherForecast/Enumeration/RequestType.swift
@@ -1,12 +1,0 @@
-//
-//  RequestType.swift
-//  WeatherForecast
-//
-//  Created by 홍정아 on 2021/10/01.
-//
-
-import Foundation
-
-enum RequestType {
-    case requestWithQueryItems
-}
